Bug fixes
I had to tweak some stuff. pyreadline is not sufficient for some things, so I added pyautogui's typewrite as an import. What is nice is that typewrite is cross-platform. I was also able to replace an obsolete xpath reference. There still is one left b/c I previously had issues with it. Someday, I might tackle it again.